+++ packages/fast-component-explorer/app/components/glyph/glyph.props.ts
@@ -1,0 +1,12 @@
+export enum Icon {
+    play = "play",
+    pause = "pause",
+    robot = "robot",
+    user = "user",
+    arrow = "arrow",
+    download = "download",
+}
+
+export interface GlyphProps {
+    path: Icon;
+}
