+++ packages/fast-component-explorer/app/components/group/group.props.ts
@@ -2,4 +2,6 @@ import React from "react";
 
 export interface GroupProps {
     style?: React.CSSProperties;
+    tag?: string;
+    slot?: string;
 }
