--- packages/fast-component-explorer/app/components/group/group.schema.json
@@ -1,17 +1,0 @@
-{
-    "$schema": "http://json-schema.org/schema#",
-    "title": "Group",
-    "description": "A group's schema definition.",
-    "id": "group",
-    "type": "object",
-    "properties": {},
-    "reactProperties": {
-        "children": {
-            "title": "Children",
-            "type": "children",
-            "defaults": [
-                "text"
-            ]
-        }
-    }
-}
