+++ packages/fast-component-explorer/app/components/group/group.schema.ts
