+++ packages/fast-component-explorer/app/components/index.ts
@@ -1,1 +1,2 @@
+export * from "./glyph";
 export * from "./group";
