--- packages/fast-component-explorer/app/utilities/components/react-html-element.tsx
@@ -1,30 +1,0 @@
-import React from "react";
-
-export interface ReactHTMLElementProps {
-    tag?: string;
-    slot?: string;
-}
-
-/**
- * A stand-in for any arbitrary HTML element that needs to be added for testing purposes
- */
-export default class ReactHTMLElement extends React.Component<ReactHTMLElementProps, {}> {
-    public static defaultProps: ReactHTMLElementProps = {
-        tag: "div",
-        slot: undefined,
-    };
-
-    public render(): JSX.Element {
-        const props: any = Object.assign({}, this.props);
-
-        delete props.tag;
-        delete props.slot;
-        delete props.children;
-
-        return <this.tag {...props}>{this.props.children}</this.tag>;
-    }
-
-    private get tag(): any {
-        return this.props.tag;
-    }
-}
