+++ packages/fast-component-explorer/app/utilities/configs/action-toggle.ts
@@ -5,8 +5,7 @@ import {
     ActionToggleProps,
     actionToggleSchema,
 } from "@microsoft/fast-components-react-msft";
-import glyphSchema from "../components/glyph.schema";
-import { Icon } from "../components/glyph";
+import { glyphSchema, Icon } from "../../../app/components/";
 import Guidance from "../../.tmp/action-toggle/guidance";
 
 const actionToggleConfig: ComponentViewConfig<ActionToggleProps> = {
