+++ packages/fast-component-explorer/app/utilities/configs/radio.ts
@@ -1,9 +1,9 @@
 import { ComponentViewConfig } from "./data.props";
 import {
+    labelSchema,
     Radio,
     RadioProps,
     radioSchema,
-    labelSchema,
 } from "@microsoft/fast-components-react-msft";
 import { uniqueId } from "lodash-es";
 import Guidance from "../../.tmp/radio/guidance";
