+++ packages/fast-component-explorer/app/utilities/configs/select-option.ts
@@ -5,8 +5,7 @@ import {
     selectOptionSchema,
 } from "@microsoft/fast-components-react-msft";
 import { uniqueId } from "lodash-es";
-import glyphSchema from "../components/glyph.schema";
-import { Icon } from "../components/glyph";
+import { glyphSchema, Icon } from "../../../app/components/";
 import Guidance from "../../.tmp/select-option/guidance";
 
 const selectOptionConfig: ComponentViewConfig<SelectOptionProps> = {
